backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
python-dotenv==1.0.0
pymupdf==1.23.7
requests==2.31.0
numpy==1.26.2
scikit-learn==1.3.2
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
torch==2.1.0+cpu --extra-index-url https://download.pytorch.org/whl/cpu
transformers==4.36.0
nltk==3.8.1
sentence-transformers==2.2.2
python-Levenshtein==0.23.0
faiss-cpu==1.7.4
huggingface-hub==0.19.4
accelerate==0.25.0
protobuf==4.25.1